chore: refeator the code
