+++ mojo_opset/tests/accuracy/operators/test_attention_quant.py
@@ -620,12 +620,6 @@ def test_paged_decode_swa_with_kv_dequant(
 # ===========================================================================
 # MojoPagedPrefillSageGQA
 # ===========================================================================
-# NOTE: There is currently no accelerator backend for ``MojoPagedPrefillSageGQA``
-# (only the CPU/torch reference). This minimal test merely exercises the
-# python-level forward path with dynamic per-token int8 quant (query_scale /
-# key_scale / value_scale all None). Once a real backend is registered, this
-# matrix should be expanded similar to ``test_paged_prefill_gqa_with_kv_dequant``
-# above (more shapes, gqa_layout / dtype combinations).
 
 test_configs_prefill_sage_gqa = [
     (2, 16, 4, 128, 1024, 0, 32, torch.bfloat16, "M_BF16"),
@@ -676,8 +670,6 @@ def test_paged_prefill_sage_gqa(
     context_dtype: torch.dtype,
     compute_dtype: torch.dtype,
 ):
-    
-
     # Sage runs dynamic per-token int8 quant *inside* its forward, so the
     # caller still passes float K/V cache and ``None`` scales — the dtype
     # arguments only configure the operator's internal quant matmul path.
@@ -701,6 +693,9 @@ def test_paged_prefill_sage_gqa(
 
     v_cache_q, value_scale = _quantize_kv_cache(v_cache, context_dtype)
 
+    atol = 5e-2 if query.dtype != torch.float32 else 1e-5
+    rtol = 5e-2 if query.dtype != torch.float32 else 1e-6
+
     op.forward_diff_with(
         op_ref,
         query,
@@ -715,6 +710,7 @@ def test_paged_prefill_sage_gqa(
         cu_total_seq_lens=cu_total_seq_lens,
         max_q_lens=max_q_lens,
         max_total_seq_lens=max_total_seq_lens,
-        atol=2e-2 if query.dtype != torch.float32 else 1e-5,
-        rtol=2e-2 if query.dtype != torch.float32 else 1e-6,
+        atol=atol,
+        rtol=rtol,
+        ptol=0.90,
     )
